+++ src/hooks/reactQuery/logout/useLogoutMutation.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { logout } from "@/components/Auth/sign";
+import { logout } from "@/utils/sign";
 
 const fetchLogout = async () => {
   await logout();
